requirements/requirements.txt
torch==2.1.2 
numpy==1.26.3
hydra-core==1.3.2
tqdm==4.66.1
